Moving Tabs to components
