+++ packages/app/features/NetworkLayout/WalletCard/shared.ts
@@ -2,7 +2,7 @@ import { type ImageSourcePropType } from 'react-native';
 
 export interface CardSkin {
 	backgroundSrc: ImageSourcePropType;
-	largeIconSrc: ImageSourcePropType;
+	largeIconSrc?: ImageSourcePropType;
 	iconSrc: ImageSourcePropType;
 	iconSize: number;
 	iconColor: string;
